remove ast-tools package and move logic to core
