--- packages/ast-tools/README.md
@@ -1,16 +1,0 @@
-# @stylable/build-tools
-
-[![npm version](https://img.shields.io/npm/v/@stylable/build-tools.svg)](https://www.npmjs.com/package/@stylable/build-tools)
-
-A collection of helper functions for Stylable based tooling.
-
-## What's inside
-
-* processUrlDependencies - Collect all dependencies of Stylable meta and replace them with placeholder
-* hasImportedSideEffects - Detect if a meta has imports with Stylable side effects
-* loadStylableConfig - Loads `stylable.config` file from package
-* sortModulesByDepth - Sort algorithm for Stylable modules 
-
-
-## License
-Copyright (c) 2021 Wix.com Ltd. All Rights Reserved. Use of this source code is governed by an [MIT license](./LICENSE).
