--- packages/ast-tools/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@stylable/ast-tools",
-  "version": "4.7.4",
-  "description": "Collection of helper functions for Stylable ast parsing and editing.",
-  "main": "dist/index.js",
-  "scripts": {
-    "test": "mocha \"./dist/test/**/*.spec.js\" --timeout 20000"
-  },
-  "dependencies": {
-    
-  },
-  "files": [
-    "dist",
-    "!dist/test",
-    "src",
-    "!*/tsconfig.{json,tsbuildinfo}"
-  ],
-  "engines": {
-    "node": ">=12"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "https://github.com/wix/stylable/tree/master/packages/build-tools",
-  "author": "Wix.com",
-  "license": "MIT"
-}
