--- packages/ast-tools/test/index.spec.ts
@@ -1,7 +1,0 @@
-import { expect } from 'chai';
-
-describe('*', () => {
-    it('*', () => {
-        expect(false).to.eql(true);
-    });
-});
