--- packages/ast-tools/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "../dist/test",
-    "types": ["node", "externals", "mocha"]
-  },
-  "references": [{ "path": "../src" }]
-}
