+++ tsconfig.json
@@ -1,8 +1,6 @@
 {
   "files": [],
   "references": [
-    { "path": "./packages/ast-tools/test" },
-    { "path": "./packages/ast-tools/src" },
     { "path": "./packages/build-tools/test" },
     { "path": "./packages/build-tools/src" },
     { "path": "./packages/cli/src" },
